Fix requests where the client requests an upgrade but the server declines.
In these cases, the correct behaviour is to send the response back to
the client and then close the connection. The client may have sent us
bytes after the request header which are not HTTP, so we have no hope
of recovering this connection.

This is achieved by modifying the parser to stop if it detects that
the client has requested an upgrade. This ensures that, regardless of
whether or not [Reqd.input_state] is [Upgrade], the parser will not
attempt to consume bytes that might not be HTTP.